added experimental AMF3 support
git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@1596 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ src/org/red5/io/object/BaseInput.java
@@ -33,14 +33,14 @@ public class BaseInput {
 
 	protected Map refMap = new HashMap();
 
-	protected short refId;
+	protected int refId;
 
 	/**
 	 * Store an object into a map
 	 * @param obj
 	 */
 	public void storeReference(Object obj) {
-		refMap.put(Short.valueOf(refId++), obj);
+		refMap.put(Integer.valueOf(refId++), obj);
 	}
 
 	/**
@@ -56,8 +56,8 @@ public class BaseInput {
 	 * @param id
 	 * @return Object
 	 */
-	protected Object getReference(short id) {
-		return refMap.get(Short.valueOf(id));
+	protected Object getReference(int id) {
+		return refMap.get(Integer.valueOf(id));
 	}
 
 }
